VPN-6213 part 6 - rework files
