+++ nebula/ui/themes/main/theme.js
@@ -2,93 +2,107 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-const theme = {};
+/**
+ * Part 3: Theme colors
+ * Named colors in this section should only be set to colors in part 1 or 2.
+ * This file should be duplicated for additional themes, and is the only file
+ * that needs to be duplicated.
+ *
+ * The named colors in this section and section 4 SHOULD be used in code -
+ * and should be the ONLY colors used in code. (One exception: color.transparent
+ * can be used in code.)
+ */
+// Primary
+color.primary = color.purple90;
+color.primaryBrighter = color.purple70;
+color.primaryHovered = color.washedPurple;
+color.primaryPressed = color.dullPurple;
 
-theme.fontFamily = 'Metropolis';
-theme.fontBoldFamily = 'MetropolisSemiBold';
-theme.fontInterFamily = 'InterUI';
-theme.fontInterSemiBoldFamily = 'InterSemiBold';
-theme.fontSize = 15;
-theme.fontSizeLarge = 22;
-theme.fontSizeSmall = 13;
-theme.fontSizeSmallest = 11;
-theme.fontWeightBold = 600;
-theme.iconSize = 16;
-theme.iconSizeFlag = 72;
-theme.iconSizeSmall = 14;
-theme.labelLineHeight = 22;
-theme.cityListTopMargin = 18;
-theme.controllerInterLineHeight = 18;
-theme.hSpacing = 20;
-theme.vSpacing = 24;
-theme.vSpacingSmall = 16;
-theme.listSpacing = 8;
-theme.maxTextWidth = 296;
-theme.windowMargin = 16;
-theme.popupMargin = 24;
-theme.sheetTopMargin = 32
-theme.desktopAppHeight = 640;
-theme.desktopAppWidth = 360;
-theme.tabletMinimumWidth = 600;
-theme.largePhoneHeight = 852
-theme.menuHeight = 56;
-theme.viewBaseTopMargin = 16;
-theme.checkBoxRowSubLabelTopMargin = 2;
-theme.dividerHeight = 1
+// Backgrounds
+color.bgColor = color.grey5;
+color.bgColorTransparent = addTransparency(color.bgColor, 0.0);
+color.bgColorStronger = color.white;
+color.overlayBackground = addTransparency(color.black, 0.0);
 
-theme.helpSheetTitleBodySpacing = 8
-theme.helpSheetBodySpacing = 16
-theme.helpSheetBodyButtonSpacing = 16
-theme.helpSheetSecondaryButtonSpacing = 8
+// Fonts
+color.fontColor = color.grey40;
+color.fontColorDark = color.grey50;
+color.fontColorWarning = color.yellow70;
+color.fontColorInverted = color.white;
+color.fontColorInvertedMuted = color.grey20;
+color.fontTitleColor = color.black;
 
-theme.navBarHeight = 64;
-theme.navBarMaxWidth = 608;
-theme.navBarTopMargin = 48;
-theme.navBarBottomMargin = Qt.platform.os !== 'ios' ? 16 : 34;
-theme.navBarHeightWithMargins =
-theme.navBarHeight + theme.navBarTopMargin + theme.navBarBottomMargin;
-theme.navBarIconSize = 48
-theme.navBarMaxPaddingTablet = 120
-theme.navBarMaxPadding = 48
+// Dividers and borders
+color.divider = color.grey10;
+color.dropShadow = color.grey60;
+color.darkFocusBorder = color.grey40;
+color.lightFocusBorder = color.grey15;
+color.inputHighlight = color.grey10;
 
-theme.onboardingMinimumVerticalSpacing = 16;
+// Success/Failure/Warning/Normal levels
+color.successMain = color.green90;
+color.successAccent = color.green;
+color.successBackground = color.green5;
+color.errorMain = color.red80;
+color.errorAccent = color.red;
+color.errorAccentLight = color.red50;
+color.errorBackground = color.red10;
+color.warningMain = color.yellow90;
+color.warningAccent = color.orange;
+color.warningBackground = color.orange10;
+color.normalLevelMain = color.blue90;
+color.normalLevelAccent = color.blue;
+color.normalLevelBackground = addTransparency(color.blue, 0.4);
 
-theme.swipeDelegateActionWidth = 56;
-theme.badgeHorizontalPadding = 8
-theme.badgeVerticalPadding = 3
+// Button colors
+color.radioButtonDisabledHovered = color.greyHovered;
+color.radioButtonDisabledPressed = color.greyPressed;
+color.disabledButtonColor = color.grey20;
 
-theme.checkBoxHeightWidth = 20;
-theme.checkmarkHeightWidth = 12;
+// Step progress bar (several other colors for bar based on primary color)
+color.stepProgressBarHighlight = color.purple10;
+color.stepProgressBarIncomplete = color.grey30;
 
-theme.progressBarDelegateHeightWidth = 32
+// Toggle
+color.activatingToggle = color.dullPurple;
+color.activatingToggleBackground = color.dullGreen;
 
-theme.maxZLevel = 10
+// Objects (these aren't colors, these are objects of colors)
+color.destructiveButton = color.redButton;
+color.normalButton = color.blueButton;
+color.successAlert = color.greenAlert;
+color.warningAlert = color.orangeAlert;
+color.clickableRow = color.clickableRowBlue;
+color.textLink = color.greyLink;
 
-theme.toggleHeight = 24
-theme.toggleWidth = 45
-theme.toggleRowDividerSpacing = 16
-
-theme.opacityDisabled = .5;
-
-theme.cornerRadius = 4;
-theme.focusBorderWidth = 2;
-
-// In milliseconds, the animation of a single device removal
-theme.removeDeviceAnimation = 300;
-theme.rowHeight = 40;
-theme.settingsMaxContentHeight = 740;
-theme.maxHorizontalContentWidth = 460;
-theme.contentTopMarginDesktop = 20;
-theme.contentTopMarginMobile = 48;
-theme.contentBottomMargin = 20;
-
-theme.uiState = {
-  'stateDefault': 'state-default',
-  'stateHovered': 'state-hovered',
-  'statePressed': 'state-pressed',
-  'stateFocused': 'state-focused',
-  'stateLoading': 'state-loading',
-  'stateDisabled': 'state-disabled',
+color.onBoardingGradient = {
+  start: color.lighterOnboardingPurple,
+  middle: color.mediumOnboardingPurple,
+  end: color.darkerOnboardingPurple,
 };
 
-theme;
+color.vpnToggleConnected = {
+  defaultColor: color.green,
+  buttonHovered: color.green60,
+  buttonPressed: color.green70,
+  focusOutline: addTransparency(color.bgColor, 0.3),
+  focusBorder: color.lightFocusBorder,
+};
+
+color.vpnToggleDisconnected = {
+  defaultColor: color.grey30,
+  buttonHovered: color.fontColor,
+  buttonPressed: color.fontColorDark,
+  buttonDisabled: color.disabledButtonColor,
+  focusOutline: color.transparent,
+  focusBorder: color.darkFocusBorder,
+};
+
+color.card = {
+  defaultColor: color.white,
+  buttonHovered: color.greyHovered,
+  buttonPressed: color.greyPressed,
+  buttonDisabled: color.bgColorTransparent,
+  focusOutline: color.bgColorTransparent,
+  focusBorder: color.darkFocusBorder,
+};
